+++ crates/dxm-resources/src/lib.rs
@@ -1,8 +1,10 @@
 //! A crate for installing third-party resources for FXServer.
 
-use std::{error::Error, path::Path};
+use std::{
+    error::Error,
+    path::{Path, PathBuf},
+};
 
-use fs_extra::dir::{CopyOptions, move_dir};
 use reqwest::blocking::Client;
 use tempfile::TempDir;
 
@@ -39,26 +41,113 @@ where
 
     fs_err::create_dir_all(path)?;
 
-    let copy_options = CopyOptions::new().content_only(true);
-    let mut original_dir: Option<TempDir> = None;
-
-    if is_dir_with_files(path)? {
-        let tempdir = TempDir::with_suffix("dxm-resource-original")?;
-        move_dir(path, &tempdir, &copy_options)?;
-        original_dir = Some(tempdir);
-    }
-
+    let vacated_dir = VacatedDir::temp(path)?;
     let result = source.download(path, nested_path);
+
     if result.is_err()
-        && let Some(original_dir) = original_dir
+        && let Some(vacated_dir) = vacated_dir
     {
-        move_dir(original_dir, path, &copy_options)?;
+        vacated_dir.bring_back()?;
     }
 
     result?;
     fs_err::write(path.join(".gitignore"), ROOT_GITIGNORE)?;
 
     Ok(())
+}
+
+fn move_dir_contents<A, B>(from: A, to: B) -> fs_extra::error::Result<u64>
+where
+    A: AsRef<Path>,
+    B: AsRef<Path>,
+{
+    fs_extra::dir::move_dir(
+        from,
+        to,
+        &fs_extra::dir::CopyOptions::new().content_only(true),
+    )
+}
+
+/// Used to move a directory to a temporary location to be later brought back or
+/// deleted.
+pub struct VacatedDir {
+    vacated_path: PathBuf,
+    dest_path: PathBuf,
+    #[allow(dead_code)]
+    temp_dir: Option<TempDir>,
+}
+
+impl VacatedDir {
+    pub fn new<A, B>(from: A, to: B) -> Result<Option<Self>, Box<dyn Error>>
+    where
+        A: AsRef<Path>,
+        B: AsRef<Path>,
+    {
+        let from = from.as_ref();
+        let to = to.as_ref();
+
+        Ok(if !is_dir_with_files(from)? {
+            None
+        } else {
+            fs_err::create_dir_all(to)?;
+
+            log::debug!("vacating {} to {}", from.display(), to.display());
+
+            move_dir_contents(from, to)?;
+
+            Some(Self {
+                vacated_path: from.into(),
+                dest_path: to.into(),
+                temp_dir: None,
+            })
+        })
+    }
+
+    pub fn temp<P>(path: P) -> Result<Option<Self>, Box<dyn Error>>
+    where
+        P: AsRef<Path>,
+    {
+        let path = path.as_ref();
+
+        Ok(if !is_dir_with_files(path)? {
+            None
+        } else {
+            let temp_dir = TempDir::with_suffix("dxm-vacated")?;
+
+            log::debug!(
+                "vacating {} to temp dir {}",
+                path.display(),
+                temp_dir.path().display()
+            );
+
+            move_dir_contents(path, &temp_dir)?;
+
+            Some(Self {
+                vacated_path: path.into(),
+                dest_path: temp_dir.path().into(),
+                temp_dir: Some(temp_dir),
+            })
+        })
+    }
+
+    pub fn bring_back(self) -> Result<(), Box<dyn Error>> {
+        log::debug!(
+            "bringing back {} to {}",
+            self.vacated_path.display(),
+            self.dest_path.display()
+        );
+
+        if is_dir_with_files(&self.vacated_path)? {
+            // remove directory contents but keep directory itself
+            fs_err::remove_dir_all(&self.vacated_path)?;
+        }
+
+        fs_err::create_dir_all(&self.vacated_path)?;
+
+        move_dir_contents(self.dest_path, self.vacated_path)?;
+
+        Ok(())
+    }
 }
 
 fn is_dir_with_files<P>(path: P) -> std::io::Result<bool>