+++ src/commands/monitor/remove.rs
@@ -1,24 +1,19 @@
-//! Contains the command to install third-party resources for FXServer.
+//! Contains the command to remove the third-party FXServer monitor.
 
 use std::{error::Error, path::PathBuf};
 
 use clap::{Arg, ArgMatches, Command};
+use dxm_artifacts::cfx::ArtifactsPlatform;
 use dxm_manifest::lockfile::Lockfile;
 
 /// The command structure.
 pub fn cli() -> Command {
     Command::new("remove")
-        .about("Uninstall FXServer resources")
-        .arg(
-            Arg::new("name")
-                .help("The name of the resource to uninstall")
-                .index(1)
-                .required(true),
-        )
+        .about("Uninstall the third-party FXServer monitor")
         .arg(
             Arg::new("manifest-path")
                 .help("The path to a directory with a dxm manifest")
-                .index(2)
+                .index(1)
                 .value_parser(clap::value_parser!(PathBuf))
                 .default_value("."),
         )
@@ -26,7 +21,6 @@ pub fn cli() -> Command {
 
 /// The code ran when using the command.
 pub fn execute(args: &ArgMatches) -> Result<(), Box<dyn Error>> {
-    let name = args.get_one::<String>("name").expect("no name");
     let manifest_path = args
         .get_one::<PathBuf>("manifest-path")
         .expect("no manifest path");
@@ -34,18 +28,26 @@ pub fn execute(args: &ArgMatches) -> Result<(), Box<dyn Error>> {
     let (manifest_path, mut manifest) = crate::util::manifest::find(manifest_path)?;
     let mut lockfile = Lockfile::read(&manifest_path)?;
 
-    if let Some(resource) = manifest.resources.get(name) {
-        crate::util::resources::uninstall_single(&manifest_path, resource, name)?;
+    if let Some(monitor) = manifest.artifact.monitor() {
+        let platform = ArtifactsPlatform::default();
 
-        lockfile.remove_resource_url(name);
+        crate::util::artifacts::remove_monitor(
+            &manifest_path,
+            &manifest.artifact,
+            &platform,
+            monitor,
+            &mut lockfile,
+        )?;
 
-        manifest.resources.remove(name);
-        manifest.write_resources(&manifest_path)?;
+        manifest.artifact.remove_monitor();
+        lockfile.remove_monitor_url();
+
+        manifest.write_artifact(&manifest_path)?;
         lockfile.write(manifest_path)?;
 
         log::info!("successfully uninstalled resource");
     } else {
-        log::error!("resource {} is not installed", name);
+        log::error!("there is no third-party monitor is not installed");
     }
 
     Ok(())